requirements.txt
numpy
cobaya
mpi4py-mpich
camb
